crews.go

+++ crews.go
@@ -35,12 +35,12 @@ func (c *Crews) ReportHttp(s string) *Crews {
 func (c *Crews) PrimaryTest(s string) bool {
 	count := c.primary.Test([]byte(s))
 	adjustedCount := uint((float64(count) - c.p * float64(c.primary.count))/(c.q-c.p))
-	return  adjustedCount >= uint(( float64(c.primary.count)/c.numWebsites))
+	return  adjustedCount >= uint(( c.primary.count/c.numWebsites))
 }
 
 func (c *Crews) SecondaryTest(s string) bool {
 	count := c.secondary.Test([]byte(s))
 	adjustedCount := uint((float64(count) - c.p * float64(c.secondary.count))/(c.q-c.p))
-	return  adjustedCount >= uint(( float64(c.secondary.count)/c.numWebsites))
+	return  adjustedCount >= uint((c.secondary.count/c.numWebsites))
 }
